+++ fmpy/gui/simulation.py
@@ -8,19 +8,23 @@ class SimulationThread(QThread):
     progressChanged = pyqtSignal(int)
     messageChanged = pyqtSignal(str, str)
 
-    def __init__(self, filename, stopTime, solver, stepSize, relativeTolerance, outputInterval, startValues, input, output, parent=None):
+    def __init__(self, filename, fmiType, stopTime, solver, stepSize, relativeTolerance, outputInterval, startValues, applyDefaultStartValues, input, output, debugLogging, fmiLogging, parent=None):
 
         super(SimulationThread, self).__init__(parent)
 
         self.filename = filename
+        self.fmiType = fmiType
         self.stopTime = stopTime
         self.solver = solver
         self.stepSize = stepSize
         self.relativeTolerance = relativeTolerance
         self.outputInterval = outputInterval
         self.startValues = startValues
+        self.applyDefaultStartValues = applyDefaultStartValues
         self.input = input
         self.output = output
+        self.debugLogging = debugLogging
+        self.fmiLogging = fmiLogging
         self.progress = 0
         self.stopped = False
 
@@ -42,6 +46,9 @@ class SimulationThread(QThread):
 
         self.messageChanged.emit(level, message.decode('utf-8'))
 
+    def logFMICall(self, message):
+        self.messageChanged.emit('debug', message)
+
     def stepFinished(self, time, recorder):
 
         progress = int((time / self.stopTime) * 100)
@@ -62,16 +69,19 @@ class SimulationThread(QThread):
         startTime = QDateTime.currentMSecsSinceEpoch()
 
         try:
-            # TODO: pass fmi_type
             self.result = simulate_fmu(self.filename,
                                        stop_time=self.stopTime,
                                        solver=self.solver,
                                        step_size=self.stepSize,
                                        relative_tolerance=self.relativeTolerance,
                                        output_interval=self.outputInterval,
+                                       fmi_type=self.fmiType,
                                        start_values=self.startValues,
+                                       apply_default_start_values=self.applyDefaultStartValues,
                                        input=self.input,
                                        output=self.output,
+                                       debug_logging=self.debugLogging,
+                                       fmi_call_logger=self.logFMICall if self.fmiLogging else None,
                                        logger=self.logFMUMessage,
                                        step_finished=self.stepFinished)
         except Exception as e:
@@ -84,3 +94,6 @@ class SimulationThread(QThread):
             self.messageChanged.emit('info', 'Simulation stopped after %s s' % totalTime)
         else:
             self.messageChanged.emit('info', 'Simulation took %s s' % totalTime)
+
+    def logFMICall(self, message):
+        self.messageChanged.emit('debug', message)